Moved the library functions to their own crates, added timers
