+++ tools/mcrl2/crates/merc_pbes/src/permutation.rs
@@ -6,7 +6,7 @@ use std::fmt;
 use merc_utilities::MercError;
 
 #[derive(Clone, PartialEq, Eq)]
-pub(crate) struct Permutation {
+pub struct Permutation {
     /// We represent a permutation as an explicit list of (domain -> image) pairs,
     /// sorted by domain.
     mapping: Vec<(usize, usize)>,
@@ -16,7 +16,7 @@ impl Permutation {
     /// Create a permutation from a given mapping of (domain -> image) pairs. Internally
     /// sorts the mapping by domain for a unique representation. The input must be
     /// a valid permutation (so a bijection).
-    pub(crate) fn from_mapping(mut mapping: Vec<(usize, usize)>) -> Self {
+    pub fn from_mapping(mut mapping: Vec<(usize, usize)>) -> Self {
         debug_assert!(
             is_valid_permutation(&mapping),
             "Input mapping is not a valid permutation: {:?}",
@@ -39,7 +39,7 @@ impl Permutation {
     }
 
     /// Parse a permutation from a string input of the form "[0->2, 1->0, 2->1]".
-    pub(crate) fn from_mapping_notation(line: &str) -> Result<Self, MercError> {
+    pub fn from_mapping_notation(line: &str) -> Result<Self, MercError> {
         // Remove the surrounding brackets if present.
         let trimmed_input = line.trim();
         let input_no_brackets =
@@ -89,7 +89,7 @@ impl Permutation {
     }
 
     /// Parse a permutation in cycle notation, e.g., (0 2 1)(3 4).
-    pub(crate) fn from_cycle_notation(cycle_notation: &str) -> Result<Self, MercError> {
+    pub fn from_cycle_notation(cycle_notation: &str) -> Result<Self, MercError> {
         let mut mapping: Vec<(usize, usize)> = Vec::new();
 
         // Split the input into cycles by finding all '(...)' groups
@@ -134,7 +134,7 @@ impl Permutation {
     }
 
     /// Construct a new permutation by concatenating two (disjoint) permutations.
-    pub(crate) fn concat(self, other: &Permutation) -> Permutation {
+    pub fn concat(self, other: &Permutation) -> Permutation {
         debug_assert!(
             self.mapping
                 .iter()
@@ -149,7 +149,7 @@ impl Permutation {
     }
 
     /// Returns the value of the permutation at the given key.
-    pub(crate) fn value(&self, key: usize) -> usize {
+    pub fn value(&self, key: usize) -> usize {
         for (d, v) in &self.mapping {
             if *d == key {
                 return *v;
@@ -160,12 +160,12 @@ impl Permutation {
     }
 
     /// Returns an iterator over the domain of this permutation.
-    pub(crate) fn domain(&self) -> impl Iterator<Item = usize> + '_ {
+    pub fn domain(&self) -> impl Iterator<Item = usize> + '_ {
         self.mapping.iter().map(|(d, _)| *d)
     }
 
     /// Check whether this permutation is the identity permutation.
-    pub(crate) fn is_identity(&self) -> bool {
+    pub fn is_identity(&self) -> bool {
         self.mapping.iter().all(|(d, v)| d == v)
     }
 
@@ -175,13 +175,13 @@ impl Permutation {
     /// Used to reject a generator whose points fall outside the parameter range
     /// before it reaches [`crate::bsgs::DensePermutation`], which silently
     /// truncates to the group's degree.
-    pub(crate) fn max_point(&self) -> Option<usize> {
+    pub fn max_point(&self) -> Option<usize> {
         self.mapping.iter().map(|&(d, v)| d.max(v)).max()
     }
 }
 
 /// Checks whether the mapping represents a valid permutation
-pub(crate) fn is_valid_permutation(mapping: &[(usize, usize)]) -> bool {
+pub fn is_valid_permutation(mapping: &[(usize, usize)]) -> bool {
     let mut domain = HashSet::with_capacity(mapping.len());
     let mut image = HashSet::with_capacity(mapping.len());
 
@@ -271,7 +271,7 @@ impl fmt::Debug for Permutation {
 /// - (3 4)
 /// - (0 3 4)
 /// - (0 4 3)
-pub(crate) fn permutation_group(indices: Vec<usize>) -> impl Iterator<Item = Permutation> + Clone {
+pub fn permutation_group(indices: Vec<usize>) -> impl Iterator<Item = Permutation> + Clone {
     let n = indices.len();
 
     // Clone the indices for use in the closure.
@@ -285,7 +285,7 @@ pub(crate) fn permutation_group(indices: Vec<usize>) -> impl Iterator<Item = Per
 
 /// Returns the number of permutations in a given group (`n!`), saturating at
 /// `usize::MAX` instead of overflowing for large `n`.
-pub(crate) fn permutation_group_size(n: usize) -> usize {
+pub fn permutation_group_size(n: usize) -> usize {
     (1..=n).fold(1usize, |acc, factor| acc.saturating_mul(factor))
 }
