+++ tools/mcrl2/pbes/Cargo.toml
@@ -5,34 +5,24 @@ edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
 
+[[bin]]
+name = "merc-pbes"
+path = "src/main.rs"
+# Avoid a rustdoc output filename collision with the `merc_pbes` library crate.
+doc = false
+
 [dependencies]
 mcrl2.workspace = true
-duct.workspace = true
-petgraph.workspace = true
-oxidd.workspace = true
-
-merc_collections.workspace = true
-merc_explore.workspace = true
-merc_io.workspace = true
-merc_lts.workspace = true
-merc_symbolic.workspace = true
+merc_explore = { workspace = true, features = ["clap"] }
+merc_pbes.workspace = true
 merc_tools.workspace = true
 merc_unsafety.workspace = true
 merc_utilities.workspace = true
 merc_vpg = { workspace = true, features = ["clap"] }
+oxidd.workspace = true
 
 clap.workspace = true
+duct.workspace = true
 env_logger.workspace = true
-which.workspace = true
-itertools.workspace = true
 log.workspace = true
-rand.workspace = true
-rayon.workspace = true
-serde_json.workspace = true
-serde.workspace = true
-streaming-iterator.workspace = true
-
-[dev-dependencies]
-merc_syntax.workspace = true
-test-case.workspace = true
-thiserror.workspace = true
+which.workspace = true
