+++ tools/mcrl2/pbes/src/main.rs
@@ -21,26 +21,26 @@ use merc_unsafety::print_allocator_metrics;
 use merc_utilities::MercError;
 use merc_utilities::Timing;
 
-use crate::bsgs::Bsgs;
-use crate::explore_common::ParameterLayoutLPS;
-use crate::explore_common::PbesVertex;
-use crate::explore_common::check_parameter_basis;
-use crate::explore_common::explore_pbes_impl;
-use crate::explore_common::explore_pbes_parallel_impl;
-use crate::explore_common::symmetry_parameter_basis;
-use crate::explore_pbes::PbesLps;
-use crate::explore_pbes::explore_pbes;
-use crate::explore_pbes::explore_pbes_parallel;
-use crate::explore_srf::PbesSrfLps;
-use crate::explore_srf::explore_srf_pbes;
-use crate::explore_srf::explore_srf_pbes_parallel;
-use crate::explore_symbolic_srf::explore_pbes_symbolic;
-use crate::graph_symmetry::GapConfig;
-use crate::graph_symmetry::graph_symmetries;
-use crate::graph_symmetry::write_dot;
-use crate::permutation::Permutation;
-use crate::quotient_lps::QuotientLps;
-use crate::symmetry::SymmetryAlgorithm;
+use merc_pbes::Bsgs;
+use merc_pbes::GapConfig;
+use merc_pbes::ParameterLayoutLPS;
+use merc_pbes::PbesLps;
+use merc_pbes::PbesSrfLps;
+use merc_pbes::PbesVertex;
+use merc_pbes::Permutation;
+use merc_pbes::QuotientLps;
+use merc_pbes::SymmetryAlgorithm;
+use merc_pbes::check_parameter_basis;
+use merc_pbes::explore_pbes;
+use merc_pbes::explore_pbes_impl;
+use merc_pbes::explore_pbes_parallel;
+use merc_pbes::explore_pbes_parallel_impl;
+use merc_pbes::explore_pbes_symbolic;
+use merc_pbes::explore_srf_pbes;
+use merc_pbes::explore_srf_pbes_parallel;
+use merc_pbes::graph_symmetries;
+use merc_pbes::symmetry_parameter_basis;
+use merc_pbes::write_dot;
 
 use merc_explore::CacheLPS;
 use merc_explore::CachingStrategy;
@@ -54,24 +54,10 @@ use merc_vpg::solve_zielonka;
 use merc_vpg::verify_solution;
 use merc_vpg::write_pg;
 
-mod bsgs;
-mod clone_iterator;
-mod explore_common;
-mod explore_pbes;
-mod explore_pbes_test;
-mod explore_srf;
-mod explore_srf_test;
-mod explore_symbolic_srf;
-mod explore_symbolic_srf_test;
-mod graph_symmetry;
-mod permutation;
-mod quotient_lps;
-mod symmetry;
-
 /// Default number of nodes for the Oxidd LDD manager.
 const DEFAULT_OXIDD_NODE_CAPACITY: usize = 1 << 24;
 
-#[derive(clap::ValueEnum, Clone, Debug)]
+#[derive(clap::ValueEnum, Clone, Copy, Debug)]
 enum PbesFormat {
     Text,
     Pbes,
@@ -131,76 +117,27 @@ enum Commands {
     /// Explore a PBES explicitly into a parity game.
     ExploreExplicit(ExploreExplicitArgs),
     /// Explore a PBES symbolically using LDD-based reachability.
-    ExploreSymbolic(ExploreSymbolicArgs),
+    ExploreSymbolic(InputArgs),
     /// Solve a PBES by exploring it into a parity game and solving the game.
     Solve(SolveArgs),
 }
 
+/// The PBES to read, shared by every subcommand.
 #[derive(clap::Args, Debug)]
-struct SymmetryArgs {
+struct InputArgs {
     /// The input PBES file.
     filename: String,
 
     /// Explicitly choose the format of the input PBES file.
     #[arg(long, short('i'), value_enum)]
     format: Option<PbesFormat>,
-
-    /// Pass a single permutation in mapping notation '[0->1,1->0,...]' or cycles notation '(0 1)' to check whether it is a symmetry.
-    #[arg(long)]
-    permutation: Option<String>,
-
-    /// Search for all symmetries instead of only the first one.    
-    #[arg(long, default_value_t = false)]
-    all_symmetries: bool,
-
-    /// Partition data parameters into their sorts before considering their permutation groups.
-    #[arg(long, default_value_t = false)]
-    partition_data_sorts: bool,
-
-    /// Partition data parameters based on their updates.
-    #[arg(long, default_value_t = false)]
-    partition_data_updates: bool,
-
-    /// Print the symmetry in the mapping notation instead of the cycle notation.
-    #[arg(long, default_value_t = false)]
-    mapping_notation: bool,
-
-    /// Print the SRF representation of the PBES.
-    #[arg(long, default_value_t = false)]
-    print_srf: bool,
 }
 
+/// How to turn a PBES into a parity game, shared by every subcommand that
+/// explores one. `solve` is `explore-explicit` followed by solving the resulting
+/// game, so both accept exactly these flags.
 #[derive(clap::Args, Debug)]
-struct GraphSymmetryArgs {
-    /// The input PBES file.
-    filename: String,
-
-    /// Explicitly choose the format of the input PBES file.
-    #[arg(long, short('i'), value_enum)]
-    format: Option<PbesFormat>,
-
-    /// Path or name of the GAP executable.
-    #[arg(long, default_value = "gap")]
-    gap_path: String,
-
-    /// Write the generated GAP script to this file (for debugging).
-    #[arg(long)]
-    dump_gap_script: Option<String>,
-
-    /// Write the SDG as a Graphviz DOT file to this path.
-    #[arg(long)]
-    dot: Option<String>,
-
-    /// Print symmetries in mapping notation instead of cycle notation.
-    #[arg(long, default_value_t = false)]
-    mapping_notation: bool,
-}
-
-#[derive(clap::Args, Debug)]
-struct ExploreExplicitArgs {
-    /// The input PBES file.
-    filename: String,
-
+struct ExploreArgs {
     /// Strategy to explore the state space of the PBES. Only used for sequential
     /// exploration; ignored when `--threads > 1` (the parallel explorer always
     /// uses a level-synchronised BFS).
@@ -210,10 +147,6 @@ struct ExploreExplicitArgs {
     /// Caching strategy to use during exploration.
     #[arg(long, value_enum, default_value_t = CachingStrategy::None)]
     caching: CachingStrategy,
-
-    /// Explicitly choose the format of the input PBES file.
-    #[arg(long, short('i'), value_enum)]
-    format: Option<PbesFormat>,
 
     /// Number of worker threads used for exploration.
     #[arg(long, default_value_t = 1)]
@@ -250,33 +183,73 @@ struct ExploreExplicitArgs {
 }
 
 #[derive(clap::Args, Debug)]
-struct ExploreSymbolicArgs {
-    /// The input PBES file.
-    filename: String,
+struct SymmetryArgs {
+    #[command(flatten)]
+    input: InputArgs,
 
-    /// Explicitly choose the format of the input PBES file.
-    #[arg(long, short('i'), value_enum)]
-    format: Option<PbesFormat>,
+    /// Pass a single permutation in mapping notation '[0->1,1->0,...]' or cycles notation '(0 1)' to check whether it is a symmetry.
+    #[arg(long)]
+    permutation: Option<String>,
+
+    /// Search for all symmetries instead of only the first one.    
+    #[arg(long, default_value_t = false)]
+    all_symmetries: bool,
+
+    /// Partition data parameters into their sorts before considering their permutation groups.
+    #[arg(long, default_value_t = false)]
+    partition_data_sorts: bool,
+
+    /// Partition data parameters based on their updates.
+    #[arg(long, default_value_t = false)]
+    partition_data_updates: bool,
+
+    /// Print the symmetry in the mapping notation instead of the cycle notation.
+    #[arg(long, default_value_t = false)]
+    mapping_notation: bool,
+
+    /// Print the SRF representation of the PBES.
+    #[arg(long, default_value_t = false)]
+    print_srf: bool,
+}
+
+#[derive(clap::Args, Debug)]
+struct GraphSymmetryArgs {
+    #[command(flatten)]
+    input: InputArgs,
+
+    /// Path or name of the GAP executable.
+    #[arg(long, default_value = "gap")]
+    gap_path: String,
+
+    /// Write the generated GAP script to this file (for debugging).
+    #[arg(long)]
+    dump_gap_script: Option<String>,
+
+    /// Write the SDG as a Graphviz DOT file to this path.
+    #[arg(long)]
+    dot: Option<String>,
+
+    /// Print symmetries in mapping notation instead of cycle notation.
+    #[arg(long, default_value_t = false)]
+    mapping_notation: bool,
+}
+
+#[derive(clap::Args, Debug)]
+struct ExploreExplicitArgs {
+    #[command(flatten)]
+    input: InputArgs,
+
+    #[command(flatten)]
+    explore: ExploreArgs,
 }
 
 #[derive(clap::Args, Debug)]
 struct SolveArgs {
-    /// The input PBES file.
-    filename: String,
+    #[command(flatten)]
+    input: InputArgs,
 
-    /// Strategy to explore the state space of the PBES. Only used for sequential
-    /// exploration; ignored when `--threads > 1` (the parallel explorer always
-    /// uses a level-synchronised BFS).
-    #[arg(long, value_enum, default_value_t = ExplorationStrategy::Bfs)]
-    strategy: ExplorationStrategy,
-
-    /// Caching strategy to use during exploration.
-    #[arg(long, value_enum, default_value_t = CachingStrategy::None)]
-    caching: CachingStrategy,
-
-    /// Explicitly choose the format of the input PBES file.
-    #[arg(long, short('i'), value_enum)]
-    format: Option<PbesFormat>,
+    #[command(flatten)]
+    explore: ExploreArgs,
 
     /// Sets the algorithm used to solve the resulting parity game.
     #[arg(long, value_enum, default_value_t = Solver::Zielonka)]
@@ -285,35 +258,6 @@ struct SolveArgs {
     /// Whether to verify the solution after computing it.
     #[arg(long, default_value_t = false)]
     verify_solution: bool,
-
-    /// Number of worker threads used for exploration.
-    #[arg(long, default_value_t = 1)]
-    threads: usize,
-
-    /// Pin each worker thread round-robin to the available CPU cores.
-    #[arg(long, default_value_t = false)]
-    pinned: bool,
-
-    /// Apply symmetry reduction: compute graph automorphisms, build a BSGS, and
-    /// canonicalize every next-state to its orbit representative before adding
-    /// it to the state space.
-    #[arg(long, default_value_t = false)]
-    symmetry: bool,
-
-    /// Supply generators directly in mapping '[0->1,...]' or cycle '(0 1)'
-    /// notation to build the BSGS without running GAP symmetry detection.
-    /// Repeat the flag for multiple generators: `--quotient '(0 1)' --quotient '(2 3)'`.
-    #[arg(long, value_name = "PERM")]
-    quotient: Vec<String>,
-
-    /// Path or name of the GAP executable used to compute the BSGS (only
-    /// relevant when `--symmetry` or `--quotient` is set).
-    #[arg(long, default_value = "gap")]
-    gap_path: String,
-
-    /// Convert to SRF before exploring (legacy; default is the direct structure-graph algorithm).
-    #[arg(long, default_value_t = false)]
-    srf: bool,
 }
 
 fn main() -> ExitCode {
@@ -348,97 +292,105 @@ fn handle_command(cli: &Cli, timing: &Timing) -> Result<(), MercError> {
 
     if let Some(command) = &cli.commands {
         match command {
-            Commands::Symmetry(args) => handle_symmetry(args, preprocess)?,
-            Commands::GraphSymmetry(args) => handle_graph_symmetry(args, preprocess)?,
-            Commands::ExploreExplicit(args) => handle_explore_explicit(args, preprocess)?,
+            Commands::Symmetry(args) => handle_symmetry(args, timing, preprocess)?,
+            Commands::GraphSymmetry(args) => handle_graph_symmetry(args, timing, preprocess)?,
+            Commands::ExploreExplicit(args) => handle_explore_explicit(args, timing, preprocess)?,
             Commands::ExploreSymbolic(args) => handle_explore_symbolic(cli, args, timing, preprocess)?,
-            Commands::Solve(args) => handle_solve(args, preprocess)?,
+            Commands::Solve(args) => handle_solve(args, timing, preprocess)?,
         }
     }
 
     Ok(())
 }
 
-/// Reads a PBES from the given file in the explicitly chosen format, or the
-/// binary PBES format when no format is given.
-///
-/// Unless `preprocess` is false, the PBES is put through the same preprocessing
-/// that mCRL2 applies before instantiating one. Doing it here rather than inside
-/// a single explorer keeps every consumer of this PBES — the explorers, the
-/// symmetry detection and the parameter basis the generators index into —
-/// looking at the same equations.
-fn read_pbes(filename: &str, format: Option<PbesFormat>, preprocess: bool) -> Result<Pbes, MercError> {
-    let mut pbes = match format.unwrap_or(PbesFormat::Pbes) {
-        PbesFormat::Pbes => Pbes::from_file(filename)?,
-        PbesFormat::Text => Pbes::from_text_file(filename)?,
-    };
+impl InputArgs {
+    /// Reads the PBES in the explicitly chosen format, or the binary PBES format
+    /// when no format is given.
+    ///
+    /// Unless `preprocess` is false, the PBES is put through the same preprocessing
+    /// that mCRL2 applies before instantiating one. Doing it here rather than inside
+    /// a single explorer keeps every consumer of this PBES — the explorers, the
+    /// symmetry detection and the parameter basis the generators index into —
+    /// looking at the same equations.
+    fn read(&self, timing: &Timing, preprocess: bool) -> Result<Pbes, MercError> {
+        let mut pbes = timing.measure("load", || match self.format.unwrap_or(PbesFormat::Pbes) {
+            PbesFormat::Pbes => Pbes::from_file(&self.filename),
+            PbesFormat::Text => Pbes::from_text_file(&self.filename),
+        })?;
 
-    if preprocess {
-        pbes.preprocess()?;
-    } else {
-        info!("Skipping PBES preprocessing (--no-preprocess)");
+        if preprocess {
+            pbes.preprocess(timing)?;
+        } else {
+            info!("Skipping PBES preprocessing (--no-preprocess)");
+        }
+
+        Ok(pbes)
     }
-
-    Ok(pbes)
 }
 
-fn handle_explore_explicit(args: &ExploreExplicitArgs, preprocess: bool) -> Result<(), MercError> {
-    let pbes = read_pbes(&args.filename, args.format.clone(), preprocess)?;
-    let game = if !args.quotient.is_empty() {
-        let bsgs = build_bsgs_from_user_generators(&pbes, &args.quotient, &args.gap_path)?;
-        explore_with_symmetry(
-            &pbes,
-            args.srf,
-            args.strategy,
-            args.caching,
-            args.threads,
-            args.pinned,
-            bsgs,
-        )?
-    } else if args.symmetry {
-        let bsgs = build_bsgs_for_pbes(&pbes, &args.gap_path)?;
-        explore_with_symmetry(
-            &pbes,
-            args.srf,
-            args.strategy,
-            args.caching,
-            args.threads,
-            args.pinned,
-            bsgs,
-        )?
-    } else if args.threads > 1 && args.srf {
-        explore_srf_pbes_parallel(&pbes, args.threads, args.caching, args.pinned)?
-    } else if args.threads > 1 {
-        explore_pbes_parallel(pbes, args.threads, args.caching, args.pinned)?
-    } else if args.srf {
-        explore_srf_pbes(&pbes, args.strategy, args.caching)?
-    } else {
-        explore_pbes(pbes, args.strategy, args.caching)?
-    };
+impl ExploreArgs {
+    /// Explores `pbes` into a parity game, applying symmetry reduction when
+    /// generators are supplied or detected, and writes the game to `--output`.
+    fn explore(&self, pbes: Pbes, timing: &Timing) -> Result<merc_vpg::ParityGame, MercError> {
+        // Explicit generators take precedence over detection: giving both means
+        // the user already knows the group and only detection would be redundant.
+        let bsgs = if !self.quotient.is_empty() {
+            Some(build_bsgs_from_user_generators(
+                &pbes,
+                &self.quotient,
+                &self.gap_path,
+                timing,
+            )?)
+        } else if self.symmetry {
+            Some(build_bsgs_for_pbes(&pbes, &self.gap_path, timing)?)
+        } else {
+            None
+        };
+
+        let game = if let Some(bsgs) = bsgs {
+            explore_with_symmetry(&pbes, self, bsgs, timing)?
+        } else if self.threads > 1 && self.srf {
+            explore_srf_pbes_parallel(&pbes, self.threads, self.caching, self.pinned, timing)?
+        } else if self.threads > 1 {
+            explore_pbes_parallel(pbes, self.threads, self.caching, self.pinned, timing)?
+        } else if self.srf {
+            explore_srf_pbes(&pbes, self.strategy, self.caching, timing)?
+        } else {
+            explore_pbes(pbes, self.strategy, self.caching, timing)?
+        };
+
+        info!(
+            "Parity game: {} vertices, {} edges",
+            game.num_of_vertices(),
+            game.num_of_edges()
+        );
+
+        if let Some(output) = &self.output {
+            let mut output_file = File::create(output)?;
+            write_pg(&mut output_file, &game)?;
+            info!("Parity game written to '{}'", output.display());
+        }
+
+        Ok(game)
+    }
+}
+
+fn handle_explore_explicit(args: &ExploreExplicitArgs, timing: &Timing, preprocess: bool) -> Result<(), MercError> {
+    let game = args.explore.explore(args.input.read(timing, preprocess)?, timing)?;
+
     println!(
         "Parity game: {} vertices, {} edges",
         game.num_of_vertices(),
         game.num_of_edges()
     );
 
-    if let Some(output) = &args.output {
-        let mut output_file = File::create(output)?;
-        write_pg(&mut output_file, &game)?;
-        info!("Parity game written to '{}'", output.display());
-    }
-
     Ok(())
 }
 
 /// Handles symbolic exploration of a PBES, reporting the number of reachable
 /// BES equations (states).
-fn handle_explore_symbolic(
-    cli: &Cli,
-    args: &ExploreSymbolicArgs,
-    timing: &Timing,
-    preprocess: bool,
-) -> Result<(), MercError> {
-    let pbes = read_pbes(&args.filename, args.format.clone(), preprocess)?;
+fn handle_explore_symbolic(cli: &Cli, args: &InputArgs, timing: &Timing, preprocess: bool) -> Result<(), MercError> {
+    let pbes = args.read(timing, preprocess)?;
     let storage = init_ldd_manager(cli);
     let states = explore_pbes_symbolic(&storage, &pbes, timing)?;
     println!("Number of states: {}", states.len());
@@ -460,7 +412,12 @@ fn parse_generators(strs: &[String]) -> Result<Vec<Permutation>, MercError> {
 }
 
 /// Build a BSGS from user-supplied generator strings without running graph-symmetry detection.
-fn build_bsgs_from_user_generators(pbes: &Pbes, strs: &[String], gap_path: &str) -> Result<Arc<Bsgs>, MercError> {
+fn build_bsgs_from_user_generators(
+    pbes: &Pbes,
+    strs: &[String],
+    gap_path: &str,
+    timing: &Timing,
+) -> Result<Arc<Bsgs>, MercError> {
     let config = GapConfig {
         executable: gap_path.to_string(),
         dump_script: None,
@@ -483,7 +440,7 @@ fn build_bsgs_from_user_generators(pbes: &Pbes, strs: &[String], gap_path: &str)
         }
     }
 
-    let bsgs = Arc::new(Bsgs::from_generators(&generators, n, &config)?);
+    let bsgs = Arc::new(timing.measure("symmetry: BSGS", || Bsgs::from_generators(&generators, n, &config))?);
     info!(
         "User-supplied generators: |G| = {} ({} generator(s))",
         bsgs.order(),
@@ -493,14 +450,16 @@ fn build_bsgs_from_user_generators(pbes: &Pbes, strs: &[String], gap_path: &str)
 }
 
 /// Compute graph symmetries for `pbes` and build a BSGS from them.
-fn build_bsgs_for_pbes(pbes: &Pbes, gap_path: &str) -> Result<Arc<Bsgs>, MercError> {
+fn build_bsgs_for_pbes(pbes: &Pbes, gap_path: &str, timing: &Timing) -> Result<Arc<Bsgs>, MercError> {
     let config = GapConfig {
         executable: gap_path.to_string(),
         dump_script: None,
     };
-    let sym_result = graph_symmetries(pbes, &config)?;
+    let sym_result = timing.measure("symmetry: detection", || graph_symmetries(pbes, &config))?;
     let n = symmetry_parameter_basis(pbes)?.len();
-    let bsgs = Arc::new(Bsgs::from_generators(&sym_result.generators, n, &config)?);
+    let bsgs = Arc::new(timing.measure("symmetry: BSGS", || {
+        Bsgs::from_generators(&sym_result.generators, n, &config)
+    })?);
     info!("|G| = {} ({} generator(s))", bsgs.order(), sym_result.generators.len());
 
     // The two orders are computed by entirely separate routes — GAP's
@@ -524,12 +483,9 @@ fn build_bsgs_for_pbes(pbes: &Pbes, gap_path: &str) -> Result<Arc<Bsgs>, MercErr
 /// Explore `pbes` into a parity game, canonicalizing every next-state via `bsgs`.
 fn explore_with_symmetry(
     pbes: &Pbes,
-    srf: bool,
-    strategy: ExplorationStrategy,
-    caching: CachingStrategy,
-    threads: usize,
-    pinned: bool,
+    args: &ExploreArgs,
     bsgs: Arc<Bsgs>,
+    timing: &Timing,
 ) -> Result<merc_vpg::ParityGame, MercError> {
     // Both explorers unify with the same flags as `symmetry_parameter_basis`, so
     // they are expected to agree with it; SRF normalisation is the one that could
@@ -537,51 +493,47 @@ fn explore_with_symmetry(
     // of its own. Checking both keeps the guarantee where it can be seen.
     let basis = symmetry_parameter_basis(pbes)?;
 
-    if srf {
+    if args.srf {
         let lps = PbesSrfLps::new(pbes)?;
         check_parameter_basis(&basis, &lps.parameters(), "SRF")?;
-        quotient_explore(&lps, strategy, caching, threads, pinned, bsgs)
+        quotient_explore(&lps, args, bsgs, timing)
     } else {
         let lps = PbesLps::new(pbes.clone())?;
         check_parameter_basis(&basis, &lps.parameters(), "structure-graph")?;
-        quotient_explore(&lps, strategy, caching, threads, pinned, bsgs)
+        quotient_explore(&lps, args, bsgs, timing)
     }
 }
 
 /// Explore `lps` into a parity game, canonicalizing every next-state via `bsgs`.
 fn quotient_explore<P>(
     lps: &P,
-    strategy: ExplorationStrategy,
-    caching: CachingStrategy,
-    threads: usize,
-    pinned: bool,
+    args: &ExploreArgs,
     bsgs: Arc<Bsgs>,
+    timing: &Timing,
 ) -> Result<merc_vpg::ParityGame, MercError>
 where
     P: ParameterLayoutLPS<Value = usize, Label = (), StateInfo = PbesVertex> + Sync,
     <P::Summand as Summand>::Context: Send,
 {
-    let timing = Timing::new();
-
-    match caching {
+    match args.caching {
         CachingStrategy::None => {
             let qlps = QuotientLps::new(lps, bsgs, 1);
-            if threads > 1 {
-                explore_pbes_parallel_impl(&qlps, threads, pinned)
+            if args.threads > 1 {
+                explore_pbes_parallel_impl(&qlps, args.threads, args.pinned, timing)
             } else {
-                explore_pbes_impl(&qlps, strategy, &timing)
+                explore_pbes_impl(&qlps, args.strategy, timing)
             }
         }
-        _ => {
+        caching => {
             // The cache sits *inside* the quotient (see [`QuotientLps`]) so the
             // keys stay the narrow read-position projections of the raw states
             // instead of covering every parameter touched by canonicalization.
             let cached = CacheLPS::new(lps, caching);
             let qlps = QuotientLps::new(&cached, bsgs, 1);
-            let game = if threads > 1 {
-                explore_pbes_parallel_impl(&qlps, threads, pinned)
+            let game = if args.threads > 1 {
+                explore_pbes_parallel_impl(&qlps, args.threads, args.pinned, timing)
             } else {
-                explore_pbes_impl(&qlps, strategy, &timing)
+                explore_pbes_impl(&qlps, args.strategy, timing)
             }?;
             debug!("{}", cached.metrics());
             Ok(game)
@@ -591,49 +543,13 @@ where
 
 /// Handles the solve command, which explores a PBES into a parity game and
 /// solves the game, printing the solution of the initial vertex.
-fn handle_solve(args: &SolveArgs, preprocess: bool) -> Result<(), MercError> {
-    let pbes = read_pbes(&args.filename, args.format.clone(), preprocess)?;
-    let game = if !args.quotient.is_empty() {
-        let bsgs = build_bsgs_from_user_generators(&pbes, &args.quotient, &args.gap_path)?;
-        explore_with_symmetry(
-            &pbes,
-            args.srf,
-            args.strategy,
-            args.caching,
-            args.threads,
-            args.pinned,
-            bsgs,
-        )?
-    } else if args.symmetry {
-        let bsgs = build_bsgs_for_pbes(&pbes, &args.gap_path)?;
-        explore_with_symmetry(
-            &pbes,
-            args.srf,
-            args.strategy,
-            args.caching,
-            args.threads,
-            args.pinned,
-            bsgs,
-        )?
-    } else if args.threads > 1 && args.srf {
-        explore_srf_pbes_parallel(&pbes, args.threads, args.caching, args.pinned)?
-    } else if args.threads > 1 {
-        explore_pbes_parallel(pbes, args.threads, args.caching, args.pinned)?
-    } else if args.srf {
-        explore_srf_pbes(&pbes, args.strategy, args.caching)?
-    } else {
-        explore_pbes(pbes, args.strategy, args.caching)?
-    };
-    info!(
-        "Parity game: {} vertices, {} edges",
-        game.num_of_vertices(),
-        game.num_of_edges()
-    );
+fn handle_solve(args: &SolveArgs, timing: &Timing, preprocess: bool) -> Result<(), MercError> {
+    let game = args.explore.explore(args.input.read(timing, preprocess)?, timing)?;
 
-    let (solution, strategy) = match args.solver {
+    let (solution, strategy) = timing.measure("solve", || match args.solver {
         Solver::Zielonka => solve_zielonka(&game, args.verify_solution),
         Solver::PriorityPromotion => solve_priority_promotion(&game, args.verify_solution),
-    };
+    });
 
     if let Some(strategy) = strategy
         && args.verify_solution
@@ -651,15 +567,15 @@ fn handle_solve(args: &SolveArgs, preprocess: bool) -> Result<(), MercError> {
     Ok(())
 }
 
-fn handle_graph_symmetry(args: &GraphSymmetryArgs, preprocess: bool) -> Result<(), MercError> {
-    let pbes = read_pbes(&args.filename, args.format.clone(), preprocess)?;
+fn handle_graph_symmetry(args: &GraphSymmetryArgs, timing: &Timing, preprocess: bool) -> Result<(), MercError> {
+    let pbes = args.input.read(timing, preprocess)?;
 
     let config = GapConfig {
         executable: args.gap_path.clone(),
         dump_script: args.dump_gap_script.as_deref().map(Path::new).map(|p| p.to_path_buf()),
     };
 
-    let result = graph_symmetries(&pbes, &config)?;
+    let result = timing.measure("symmetry: detection", || graph_symmetries(&pbes, &config))?;
 
     if let Some(dot_path) = &args.dot {
         let mut f = File::create(dot_path)?;
@@ -686,8 +602,8 @@ fn handle_graph_symmetry(args: &GraphSymmetryArgs, preprocess: bool) -> Result<(
     Ok(())
 }
 
-fn handle_symmetry(args: &SymmetryArgs, preprocess: bool) -> Result<(), MercError> {
-    let pbes = read_pbes(&args.filename, args.format.clone(), preprocess)?;
+fn handle_symmetry(args: &SymmetryArgs, timing: &Timing, preprocess: bool) -> Result<(), MercError> {
+    let pbes = args.input.read(timing, preprocess)?;
     let algorithm = SymmetryAlgorithm::new(&pbes, args.print_srf)?;
     if let Some(permutation) = &args.permutation {
         let pi = if permutation.trim_start().starts_with("[") {
